% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krr_fit.R
\name{krr}
\alias{krr}
\title{Fit KRR Model}
\usage{
krr(x, y, rho = 1, lambda = 1e-04)
}
\arguments{
\item{x}{Matrix or data frame. training data.}

\item{y}{Vector. The response data.}

\item{rho}{Numeric. Parameter for the Gaussian kernel.}

\item{lambda}{Numeric. Regularization parameter.}
}
\value{
An object of class 'krr' containing the estimated coefficients.
}
\description{
Fits a Kernel Ridge Regression (KRR) model to the training data.
}
