% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krr_predict.R
\name{predict.krr}
\alias{predict.krr}
\title{Predict method for KRR Model}
\usage{
\method{predict}{krr}(model, xnew, ...)
}
\arguments{
\item{model}{An object of class 'krr'.}

\item{xnew}{Matrix or data frame. New input data for prediction.}
}
\value{
A vector (or matrix) of predicted values
}
\description{
Computes predicted values for new data using a fitted KRR model.
}
